Store member data in calendar tags.
This is not visible in the google calendar UI, but
allows us to store data in a more structured fashion, which
lets us fit the member number and email address into the event.

Once current term data is using this method, we can
stop storing data in the event description / title.